flask-server/requirements.txt
autopep8>=2.0.2
certifi>=2023.5.7
charset-normalizer>=3.2.0
click>=8.1.4
# dotenv
et-xmlfile>=1.1.0
Flask>=2.3.2
Flask-Cors>=4.0.0
idna>=3.4
itsdangerous>=2.1.2
Jinja2>=3.1.2
langchain>=0.0.227
MarkupSafe>=2.1.3
numpy>=1.25.0
openai>=0.27.8
openpyxl>=3.1.2
pandas>=2.0.3
pandas-stubs>=2.0.2.230605
pycodestyle>=2.10.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pytz>=2023.3
requests>=2.31.0
six>=1.16.0
toml>=0.10.2
tqdm>=4.65.0
urllib3>=2.0.3
Werkzeug>=2.3.6
